+++ kitchenpos-core/src/test/java/kitchenpos/application/fixture/MenuServiceFixture.java
@@ -1,16 +1,16 @@
 package kitchenpos.application.fixture;
 
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.product.domain.Product;
-import kitchenpos.common.vo.Price;
-import kitchenpos.menugroup.persistence.JpaMenuGroupRepository;
-import kitchenpos.menu.persistence.JpaMenuProductRepository;
-import kitchenpos.menu.persistence.JpaMenuRepository;
-import kitchenpos.product.persistence.JpaProductRepository;
 import kitchenpos.common.dto.request.CreateMenuRequest;
 import kitchenpos.common.dto.request.MenuProductDto;
+import kitchenpos.common.vo.Price;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menu.domain.Product;
+import kitchenpos.menu.persistence.JpaMenuProductRepository;
+import kitchenpos.menu.persistence.JpaMenuRepository;
+import kitchenpos.menu.persistence.JpaProductRepository;
+import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.persistence.JpaMenuGroupRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import javax.persistence.EntityManager;
