+++ kitchenpos-web/src/main/java/kitchenpos/ui/ProductRestController.java
@@ -1,9 +1,9 @@
 package kitchenpos.ui;
 
-import kitchenpos.product.application.ProductService;
-import kitchenpos.product.domain.Product;
 import kitchenpos.common.dto.request.CreateProductRequest;
 import kitchenpos.common.dto.response.ProductResponse;
+import kitchenpos.menu.application.ProductService;
+import kitchenpos.menu.domain.Product;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -12,7 +12,6 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.net.URI;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @RestController
 public class ProductRestController {
@@ -35,7 +34,7 @@ public class ProductRestController {
         final List<ProductResponse> productResponses = productService.list()
                                                                      .stream()
                                                                      .map(this::toResponse)
-                                                                     .collect(Collectors.toUnmodifiableList());
+                                                                     .toList();
         return ResponseEntity.ok()
                              .body(productResponses);
     }
